More fixes for Splunk certification
