+++ DA-ESS-CbResponse/README/alert_actions.conf.spec
@@ -1,0 +1,14 @@
+[isolatesensor]
+param.ipaddress       = <string>
+param.inputtype       = <string>
+param.dryrun          = <bool>
+param._cam            = json
+
+[banhash]
+param.fieldname       = <string>
+param.dryrun          = <bool>
+param._cam            = json
+
+[killprocess]
+param.dryrun          = <bool>
+param._cam            = json
