+++ DA-ESS-CbResponse/README/savedsearches.conf.spec
@@ -1,0 +1,4 @@
+param.ipaddress       = <string>
+param.inputtype       = <string>
+param.dryrun          = <bool>
+param.fieldname       = <string>
