+++ DA-ESS-CbResponse/appserver/static/setup.css
@@ -165,21 +165,6 @@ table tbody tr:hover{
     color: #fff;
 }
 
-.dialog {
-    position: fixed;
-    top: 0;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    margin: auto;
-    width: 400px;
-    height: 210px;
-    z-index: 1000;
-    background-color: #fff;
-    border-radius: 5px;
-    opacity: 1;
-    display: none;
-}
 
 .dialog-header{
     font-size: 14px;
@@ -220,37 +205,7 @@ table tbody tr:hover{
     cursor:pointer;
 }
 
-.credform input#passwordCredDialogBtnSave {
-    background-color: #65a637;
-    border: 1px solid #4e802a;
-    border-bottom-color: #4e802a;
-    border-top-color: #4e802a;
-    box-shadow: inset 0px 1px 0 #6aae3a;
-    text-shadow: none;
-    color: #ffffff;
-    border-radius: 4px;
-    padding: 2px 2px;
-    font-size: 12px;
-    float: right;
-    width: 80px;
-    height: 25px;
-}
 
-.credform input#passwordCredDialogBtnCancel {
-    background-color: #eeeeee;
-    border: 1px solid #bfbfbf;
-    border-top-color: #bfbfbf;
-    border-bottom-color: #bfbfbf;
-    color: #333333;
-    box-shadow: inset 0px 1px 0 #fdfdfd;
-    background-position: 0;
-    text-shadow: none;
-    border-radius: 4px;
-    padding: 2px 2px;
-    float: left;
-    width: 80px;
-    height: 25px;
-}
 
 .shadow{
     width: 100%;
