+++ DA-ESS-CbResponse/default/app.conf
@@ -1,17 +1,18 @@
 [install]
 state = enabled
 is_configured = false
-build = 20161012
+build = 20161017
 
 [launcher]
 author = Carbon Black Developer Network
-version = 2.0.1
+version = 2.0.2
 description = Add-on to connect Cb Response to Splunk. Ingest feed, watchlist, and raw endpoint events into Splunk to correlate across other security solutions and take action through Cb Response's Live Response capability.
 
 [ui]
 is_visible = 1
 label = Cb Response
 docs_section_override = AddOns:released
+setup_view = setup_page
 
 [package]
 id = DA-ESS-CbResponse
