+++ DA-ESS-CbResponse/default/setup.xml
@@ -1,14 +1,10 @@
-<setup>
-    <block title=" " endpoint="DA-ESS-CbResponse_input_setup/DA-ESS-CbResponse_settings" entity="DA-ESS-CbResponse_settings">
-        <input field="all_settings">
-            <label> </label>
-            <type>text</type>
-        </input>
-        <text>
-            <![CDATA[
-                <script type="text/javascript" src="/en-US/static/app/DA-ESS-CbResponse/js/setup.js"></script>
-                <script type="text/javascript" src="/en-US/static/app/DA-ESS-CbResponse/js/setup_page.js"></script>
-            ]]>
-        </text>
-    </block>
-</setup>
+<dashboard script="setup.js" stylesheet="setup.css" hideAppBar="true" hideEdit="true" isDashboard="false">
+  <label>DA-ESS-CbResponse</label>
+  <row>
+    <panel>
+      <html>
+        <div id="setup_page_container"/>
+      </html>
+    </panel>
+  </row>
+</dashboard>
