+++ README.md
@@ -199,12 +199,14 @@ these links at event generation time (see the Cb Event Forwarder configuration f
 * Deep Link to Binary Analysis page
 * Deep Link to Sensor page
 
-## Performance
+## Performance & Data Models
 
 This app contains one Data Model, representing Cb alerts plus watchlist/feed hits. 
-This data model is accelerated by default.
+The Data Model, `CbR_Alert`, is generated by searching for Cb Response 
+events with the query `tag=alert`. This data model is accelerated by default.
 
-None of the saved searches included in this app are scheduled to run by default.
+In addition, the saved search `CbResponse Alert Activity` is scheduled
+to run once per day by default.
 
 ## Support
